Added - implementation to DeleteEstateCommandHandler - Only user can delete own estate. Change Folder name in Controllers from User to Auth

+++ RealEstate.Application/Estates/Commands/DeleteEstate/DeleteEstateCommandHandler.cs
@@ -1,31 +1,49 @@
 ﻿using MediatR;
+using Microsoft.AspNetCore.Http;
+using Microsoft.AspNetCore.Identity;
 using Microsoft.EntityFrameworkCore;
 using RealEstate.Application.Common.Interfaces;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
+using RealEstate.Domain.Entities;
+using System.Security.Claims;
+using System.Xml.Linq;
 
 namespace RealEstate.Application.Estates.Commands.DeleteEstate
 {
     public class DeleteEstateCommandHandler : IRequestHandler<DeleteEstateCommand>
     {
         private readonly IEstateDbContext _context;
-        public DeleteEstateCommandHandler(IEstateDbContext estateDbContext)
+        private readonly UserManager<ApplicationUser> _userManager;
+        private readonly IHttpContextAccessor _httpContextAccessor;
+
+        public DeleteEstateCommandHandler(IEstateDbContext estateDbContext, UserManager<ApplicationUser> userManager, IHttpContextAccessor httpContextAccessor)
         {
             _context = estateDbContext;
+            _userManager = userManager;
+            _httpContextAccessor = httpContextAccessor;
         }
 
         public async Task<Unit> Handle(DeleteEstateCommand request, CancellationToken cancellationToken)
         {
+            var httpContext = _httpContextAccessor.HttpContext;
+
+            var userName = httpContext.User.FindFirstValue(ClaimTypes.Name);
+
+            var user = await _userManager.Users.FirstOrDefaultAsync(x => x.UserName == userName);
+
             var estate = await _context.Estates.Where(p => p.Id == request.EstateId).FirstOrDefaultAsync();
 
-            _context.Estates.Remove(estate);
+            if (user.Id == estate.ApplicationUserId)
+            {
+                _context.Estates.Remove(estate);
 
-            await _context.SaveChangesAsync(cancellationToken);
+                await _context.SaveChangesAsync(cancellationToken);
 
-            return Unit.Value;
+                return Unit.Value;
+            }
+            else
+            {
+                throw new Exception("This Estate doen not belong to You");
+            }
         }
     }
 }
